refactor(cli): allow running in node projects
